Module webc::compat

👎Deprecated since 5.1.0: Import types directly from the crate

A compatibility layer for dealing with different versions of the WEBC binary format.

Structs

Container

A version-agnostic read-only WEBC container.

SharedBytes

A cheaply cloneable owned buffer that may be backed by a memory-mapped file or Bytes in memory.

Volume

A WEBC volume.

Enums

ContainerError

Various errors that may occur during Container operations.

Metadata

Metadata describing the properties of a file or directory.

VolumeError

Errors that may occur when doing Volume operations.