Module webc::metadata

Package metadata.

Modules

annotations

Strongly-typed definitions for the various free-form “annotations” fields in a crate::metadata::Manifest.

Structs

Atom

Executable file, stored in the .atoms section of the file

AtomWithoutSignature

Executable file, stored in the .atoms section of the file

Binding

Command

Command that can be run by a given implementation

Manifest

Manifest of the file, see spec 2.3.1

ManifestWithoutAtomSignatures

Same as Manifest, but doesn’t require the atom.signature

WaiBindings

WitBindings

Enums

AtomSignature

BindingsExtended

UrlOrManifest

Dependency of this file, encoded as either a URL or a manifest (in case of vendored dependencies)

Type Aliases

Annotation

Annotation = free-form metadata to be used by the atom