Module webc::metadata::annotations

Strongly-typed definitions for the various free-form “annotations” fields in a crate::metadata::Manifest.

Structs

Atom

A command annotation which specifies which atom implements a command.

Emscripten

Annotations used by Emscripten runners.

FileSystemMapping

An entry used when constructing a filesystem based on a package and its dependency tree.

FileSystemMappings

A list of entries used when constructing a filesystem based on a package and its dependency tree.

VolumeSpecificPath

The path for an item inside a particular volume.

Wapm

Package-level annotations specific to the WebAssembly Package Manager.

Wasi

Annotations used by WASI runners.

Wcgi

Annotations used by WCGI runners.

Constants

EMSCRIPTEN_RUNNER_URI

The base URI used by an Emscripten runner.

WASI_RUNNER_URI

The base URI used by a Wasi runner.

WASM4_RUNNER_URI

The base URI used by a WASM4 runner.

WCGI_RUNNER_URI

The base URI used by a Wcgi runner.