Type Alias webc::metadata::Annotation

pub type Annotation = Value;

Annotation = free-form metadata to be used by the atom

Aliased Type

enum Annotation {
    Null,
    Bool(bool),
    Integer(i128),
    Float(f64),
    Bytes(Vec<u8>),
    Text(String),
    Array(Vec<Value>),
    Map(BTreeMap<Value, Value>),
    Tag(u64, Box<Value>),
    // some variants omitted
}

Variants

Null

Represents the absence of a value or the value undefined.

Bool(bool)

Represents a boolean value.

Integer(i128)

Integer CBOR numbers.

The biggest value that can be represented is 2^64 - 1. While the smallest value is -2^64. Values outside this range can’t be serialized and will cause an error.

Float(f64)

Represents a floating point value.

Bytes(Vec<u8>)

Represents a byte string.

Text(String)

Represents an UTF-8 encoded string.

Array(Vec<Value>)

Represents an array of values.

Map(BTreeMap<Value, Value>)

Represents a map.

Maps are also called tables, dictionaries, hashes, or objects (in JSON). While any value can be used as a CBOR key it is better to use only one type of key in a map to avoid ambiguity. If floating point values are used as keys they are compared bit-by-bit for equality. If arrays or maps are used as keys the comparisons to establish canonical order may be slow and therefore insertion and retrieval of values will be slow too.

Tag(u64, Box<Value>)

Represents a tagged value