Function webc::migration::are_semantically_equivalent

pub fn are_semantically_equivalent(
    v2: OwnedBuffer,
    v3: OwnedBuffer,
) -> Result<(), Error>

Checks whether two webcs (one v2 and one v3) are semantically equivalent. For v2 and v3 to be equivalent, these criteria must be true:

Contents of the manifest must be the same. In particular, these items:

origin

atoms

commands

entrypoint

Atoms in both webcs must be the same:

same number of atoms

same atom names

same atom contents

The filesystem in both webcs must be the same (regardless of structure of volumes): Each top level directory in webc v2 is equivalent to a volume in webc v3 so this check, for every top level directory in v2, will find the associated volume in v3 and walk them at the same time to make they have the same entries and contents.