Struct webc::WebCOwned

pub struct WebCOwned {
    pub webc: WebC<'static>,
    /* private fields */
}

Owned version of the WebC file that carries its data along the parsed WebC<'static>

Fields
webc: WebC<'static>
Implementations

impl WebCOwned

pub fn parse(data: Vec<u8>, options: &ParseOptions) -> ReadResult<Self>

Same as WebC::parse, but keeps the resulting data in memory, instead of referencing it

pub fn as_webc_ref<'a>(&'a self) -> WebC<'a>