Module webc::v1

Parsing code for v1 of the WEBC format.

Structs

Checksum

Calculated checksum of the file

Error

The error type used within the read module.

FsEntry

Directory or file entry, parsed without any allocation

OwnedFsEntryDir

Owned version of the FsEntry with fs_type = FsEntryType::Dir

OwnedFsEntryFile

Owned version of the FsEntry with fs_type = FsEntryType::File

ParseOptions

Options on what to parse from the file

RecursiveFsEntryDir

Signature

Signature of the checksum of the file, such that verify(WebC::get_checksum(), public_key) is valid

Volume

Filesystem volume, containing the uncompressed files in an ordered directory structure

VolumeHeader

Header of a filesystem volume, describing a serialized list of directories and file paths

VolumeIterator

Iterator over the paths in the file, yields PathBufs until all files in the volume have been listed.

WebC

Container file, lazily parsed from a set of &'data [u8] bytes

WebCMmap

Memory-mapped version of the WebC file that carries its data along the parsed WebC<'static>

WebCOwned

Owned version of the WebC file that carries its data along the parsed WebC<'static>

Enums

DirOrFile

Specifies whether an input path is a directory or a file (since this distinction can’t be made from the filename alone)

DirOrFileWithBytes

FsEntryType

Whether the file is a directory or a file

GenerateChecksum

Whether to sign the bytes when deserializing the WebC file to bytes

OwnedFsEntry

Same as FsEntry but with an owned text: String, instead of a &str

PackageInfo

Information about the package name

RecursiveFsEntry

Functions

pack_directory

webc_temp_dir

Since env::temp_dir() panics on wasm32-wasi, this function provides a non-panicking replacement

Type Definitions

FileMap

ReadResult

The result type used within the read module.