Struct webc::v1::WebC

pub struct WebC<'data> {
    pub version: u64,
    pub checksum: Option<Checksum>,
    pub signature: Option<Signature>,
    pub manifest: Manifest,
    pub atoms: Volume<'data>,
    pub volumes: IndexMap<String, Volume<'data>>,
}

Container file, lazily parsed from a set of &'data [u8] bytes

Fields
version: u64

Version of the file format

checksum: Option<Checksum>

Parsed checksum (optional in case of no encoded checksum)

signature: Option<Signature>

Parsed signature (optional if file was not signed)

manifest: Manifest

Manifest of the file, see section 2.3.1 of the spec

atoms: Volume<'data>

Executable files, indexed into one volume (a.wasm => a, b.wasm => b@0.2.1)

volumes: IndexMap<String, Volume<'data>>

Filesystem volumes: default volume name is atom (containing files of the current package) and user/package@version for external dependencies. Every dependency can be sandboxed to only access its own filesystem volume, not external ones.

Implementations

impl<'a> WebC<'a>

pub fn get_main_args_for_command( &self, command: &str ) -> Result<Vec<String>, String>

pub fn get_atom_name_for_command( &self, api: &str, command: &str ) -> Result<String, String>

pub fn check_magic_header(data: &[u8]) -> Result<(), Error>

Checks whether the file starts with the header MAGIC

pub fn get_volumes_for_package(&self, package: &str) -> Vec<String>

Determines the available volumes for a given package

pub fn list_directories(&self, volume: &str) -> Vec<String>

pub fn read_dir( &self, package: &str, path: &str ) -> Result<Vec<FsEntry<'a>>, Error>

Returns the directory entries or an error if the directory does not exist

pub fn get_file_entry( &self, package: &str, path: &str ) -> Option<(String, OwnedFsEntryFile)>

Looks for the first volume containing “entry”, scoped to the given package

pub fn get_check_version(data: &[u8]) -> Result<u64, Error>

Checks whether the version of the file is supported by the parsing implementation

pub fn get_checksum_bytes(data: &[u8]) -> Result<&[u8], Error>

Returns the bytes of the checksum

pub fn get_manifest_offset_size(data: &[u8]) -> ReadResult<(usize, usize)>

Returns the offset of the manifest start

pub fn get_manifest(data: &[u8]) -> Result<Manifest, Error>

pub fn get_atoms_volume_offset_size(data: &[u8]) -> ReadResult<(usize, usize)>

Returns the offset of the .atoms section of the file

pub fn get_atoms_volume(data: &'a [u8]) -> Result<Volume<'a>, Error>

Parses the .atoms section of the file

pub fn get_volume_data_offsets( data: &[u8] ) -> Result<BTreeMap<String, (usize, usize)>, Error>

Returns the offsets of the “volume”

pub fn parse_volumes_from_fileblock( data: &'a [u8] ) -> ReadResult<IndexMap<String, Volume<'a>>>

pub fn parse_volumes(data: &'a [u8]) -> ReadResult<IndexMap<String, Volume<'a>>>

Parses the .volumes section(s) of the file

pub fn compute_checksum(data: &[u8]) -> ReadResult<Option<Checksum>>

Computes the checksum of the file without cloning it

pub const fn get_signature_offset_start() -> usize

pub fn get_signature_offset_size(data: &[u8]) -> ReadResult<(usize, usize)>

Returns the offset of the signature

pub fn get_signature_bytes(data: &[u8]) -> ReadResult<&[u8]>

Read the signature bytes

pub fn get_signature(data: &[u8]) -> ReadResult<Option<Signature>>

Returns the (unverified) signature from the file

pub fn get_metadata(&self) -> &Manifest

Returns a reference to the manifest

pub fn get_package_name(&self) -> String

Returns the current package name with

pub fn get_atom(&self, package: &str, atom: &str) -> Result<&[u8], Error>

Returns an atom by name for a given package

pub fn get_volume(&self, package: &str, volume: &str) -> Option<&Volume<'a>>

Returns a reference to the filesystem volume of the package

pub fn get_file(&self, package: &str, file_path: &str) -> Result<&[u8], Error>

Returns a file for a given package - if you want to use a non-default volume, prefix the file_path with volume://, for example, metadata://README.md

pub fn list_volumes(&self, package: &str) -> Vec<String>

Returns a list of volumes for this package

pub fn list_packages(&self) -> Vec<PackageInfo>

Returns a list of bundled “package@version” strings contained in this package

pub fn list_atoms(&self) -> Vec<String>

Returns the atoms in the root package

pub fn get_all_atoms(&self) -> IndexMap<String, &'a [u8]>

Returns a list of all atoms with bytes

pub fn list_atoms_for_package(&self, package_orig: &str) -> Vec<String>

List the atoms for a given package

pub fn list_commands(&self) -> Vec<&str>

List the available commands for the root package

pub fn parse(data: &'a [u8], options: &ParseOptions) -> ReadResult<Self>

Parses the entire file, depending on the ParseOptions

pub fn get_volumes_as_fileblock(&self) -> Vec<u8>

pub fn into_bytes(&self, sign_bytes: GenerateChecksum) -> ReadResult<Vec<u8>>

Serialize the .webc file into bytes