Module webc::v2::read

Read files in the WEBC format.

There are several readers to choose from, depending on the use case.

An OwnedReader is the most convenient to work with because it owns the buffer containing the *.webc file, so it is able to retrieve or parse arbitrary parts of the file as necessary. Due to the use of shared_buffer::OwnedBuffer, it is possible to use mmap to avoid keeping the entire file in memory.

A StreamingReader will read data from some std::io::Read object and parse it on-demand. This isn’t as flexible as OwnedReader, but can be less memory-intensive for use cases where random access isn’t required.

Structs

AtomsSection

A section containing the atoms volume.

Directory

A directory that contains zero or more DirEntry’s.

FileEntry

The contents of a file in the volume.

IndexSection

A section containing the file’s Index.

ManifestSection

A section containing the file’s Manifest.

OwnedReader

A reader for owned data that is already in memory.

StreamingReader

A reader which can parse a WEBC file from an arbitrary Read object.

VolumeSection

A volume section containing a directory tree.

Enums

DirEntry

An item in a volume.

DirEntryError

Errors that may occur while parsing a DirEntry.

LookupError

Errors that may occur while looking up an item in a volume.

OwnedReaderError

Errors that may be emitted by OwnedReader.

Section

The different sections in a webc file.

SectionError

Errors that may occur while parsing a Section.

StreamingReaderError

Errors that may be emitted by StreamingReader.

VolumeHeaderError

Errors that may occur while reading a volume header.