Enum webc::v2::read::VolumeHeaderError

#[non_exhaustive]
pub enum VolumeHeaderError {
    NotFound,
    AccessOutOfBounds {
        offset: usize,
        header_length: usize,
    },
    InvalidFilename {
        error: Utf8Error,
        filename: Vec<u8>,
    },
    UnsupportedHeaderEntry {
        tag: Tag,
    },
    UnknownTag {
        tag: u8,
    },
    NotADirectory,
}

Errors that may occur while reading a volume header.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

NotFound

The item wasn’t found.

AccessOutOfBounds

Fields

offset: usize

header_length: usize

Tried to access something outside of the volume header.

InvalidFilename

Fields

error: Utf8Error

The underlying error.

filename: Vec<u8>

The original filename.

A filename wasn’t valid UTF-8.

UnsupportedHeaderEntry

Fields

tag: Tag

The tag that was encountered.

Encountered an entry with a tag that isn’t supported.

UnknownTag

Fields

tag: u8

The raw value.

Encountered a Tag with an known value.

NotADirectory

Found a directory when one wasn’t expected (e.g. when looking up a file’s data).