Module webc::v2::write

Create files in the WEBC format.

If you have landed on this page, you are probably looking for the Writer type.

The Writer uses The Typestate Pattern to ensure that each section of the WEBC file is written in the correct order.

Examples

use webc::{
    metadata::Manifest,
    v2::{
        write::{Writer, Directory, FileEntry},
        ChecksumAlgorithm, SignatureAlgorithm,
    },
};
use std::collections::BTreeMap;
use bytes::Bytes;

let manifest = Manifest::default();

let mut atoms = BTreeMap::new();
atoms.insert("python".parse()?, FileEntry::Borrowed(b"..."));

let mut writer = Writer::new(ChecksumAlgorithm::Sha256)
    .write_manifest(&manifest)?
    .write_atoms(atoms)?;

let volume_1 = Directory::from_path("./volume_1")?;
writer.write_volume("volume_1", volume_1)?;

let buffer: Bytes = writer.finish(SignatureAlgorithm::None)?;

Structs

Directory

A directory in a volume.

Writer

A serializer for the WEBC format.

WritingAtoms

The Writer is about to write the atoms section.

WritingManifest

The Writer is about to write the Manifest section.

WritingVolumes

The Writer is writing volumes sections.

Enums

DirEntry

A single entry in a directory.

FileEntry

Some file-like object which can be written to a WEBC file.