Struct webc::wasmer_package::Package

pub struct Package { /* private fields */ }

A Wasmer package that will be lazily loaded from disk.

Implementations

impl Package

pub fn from_tarball_file( path: impl AsRef<Path>, ) -> Result<Self, WasmerPackageError>

Load a Package from a *.tar.gz file on disk.

Implementation Details

This will unpack the tarball to a temporary directory on disk and use memory-mapped files in order to reduce RAM usage.

pub fn from_tarball_file_with_strictness( path: impl AsRef<Path>, strictness: Strictness, ) -> Result<Self, WasmerPackageError>

Load a Package from a *.tar.gz file on disk.

Implementation Details

This will unpack the tarball to a temporary directory on disk and use memory-mapped files in order to reduce RAM usage.

pub fn from_tarball(tarball: impl BufRead) -> Result<Self, WasmerPackageError>

Load a package from a *.tar.gz archive.

pub fn from_tarball_with_strictness( tarball: impl BufRead, strictness: Strictness, ) -> Result<Self, WasmerPackageError>

Load a package from a *.tar.gz archive.

pub fn from_manifest( wasmer_toml: impl AsRef<Path>, ) -> Result<Self, WasmerPackageError>

Load a package from a wasmer.toml manifest on disk.

pub fn from_manifest_with_strictness( wasmer_toml: impl AsRef<Path>, strictness: Strictness, ) -> Result<Self, WasmerPackageError>

Load a package from a wasmer.toml manifest on disk.

pub fn webc_hash(&self) -> Option<[u8; 32]>

Returns the Sha256 has of the webc represented by this Package

pub fn manifest(&self) -> &WebcManifest

Get the WEBC manifest.

pub fn atoms(&self) -> &BTreeMap<String, OwnedBuffer>

Get all atoms in this package.

pub fn volumes(&self) -> impl Iterator<Item = &Volume>

Returns all volumes in this package

pub fn serialize(&self) -> Result<Bytes, Error>

Serialize the package to a *.webc v2 file, ignoring errors due to missing files.