Crate webc

Structs

Atom

Executable file, stored in the .atoms section of the file

AtomWithoutSignature

Executable file, stored in the .atoms section of the file

Binding

Checksum

Calculated checksum of the file

Command

Command that can be run by a given implementation

Error

The error type used within the read module.

FsEntry

Directory or file entry, parsed without any allocation

IndexMap

A hash table where the iteration order of the key-value pairs is independent of the hash values of the keys.

Manifest

Manifest of the file, see spec 2.3.1

ManifestWithoutAtomSignatures

Same as Manifest, but doesn’t require the atom.signature

OwnedFsEntryDir

Owned version of the FsEntry with fs_type = FsEntryType::Dir

OwnedFsEntryFile

Owned version of the FsEntry with fs_type = FsEntryType::File

ParseOptions

Options on what to parse from the file

RecursiveFsEntryDir

Signature

Signature of the checksum of the file, such that verify(WebC::get_checksum(), public_key) is valid

Url

A parsed URL record.

Volume

Filesystem volume, containing the uncompressed files in an ordered directory structure

VolumeHeader

Header of a filesystem volume, describing a serialized list of directories and file paths

VolumeIterator

Iterator over the paths in the file, yields PathBufs until all files in the volume have been listed.

WaiBindings

WebC

Container file, lazily parsed from a set of &'data [u8] bytes

WebCOwned

Owned version of the WebC file that carries its data along the parsed WebC<'static>

WitBindings

Enums

BindingsExtended

DirOrFile

Specifies whether an input path is a directory or a file (since this distinction can’t be made from the filename alone)

DirOrFileWithBytes

FsEntryType

Whether the file is a directory or a file

GenerateChecksum

Whether to sign the bytes when deserializing the WebC file to bytes

OwnedFsEntry

Same as FsEntry but with an owned text: String, instead of a &str

PackageInfo

Information about the package name

RecursiveFsEntry

UrlOrManifest

Dependency of this file, encoded as either a URL or a manifest (in case of vendored dependencies)

Value

The Value enum, a loosely typed way of representing any valid CBOR value.

Constants

MAGIC

File identification bytes stored at the beginning of the file.

VERSION

File identification bytes stored at the beginning of the file.

Functions

pack_directory

webc_temp_dir

Since env::temp_dir() panics on wasm32-wasi, this function provides a non-panicking replacement

Type Definitions

Annotation

Annotation = free-form metadata to be used by the atom

FileMap

ReadResult

The result type used within the read module.