Crate webc

A library for reading and writing WEBC files.

The Container provides an abstraction over the various WEBC versions this crate can handle. As such, it tries to cater to the lowest common denominator and favors portability over performance or power.

In general, errors that occur during lazy operations won’t be accessible to the user.

If more flexibility is required, consider using crate::detect() and instantiating a compatible parser directly.

use webc::{
    Container,
    Version,
    v2::read::OwnedReader,
    v1::{ParseOptions, WebC},
};
let bytes: &[u8] = b"...";

match webc::detect(bytes) {
    Ok(Version::V1) => {
        let options = ParseOptions::default();
        let webc = WebC::parse(bytes, &options).unwrap();
        if let Some(signature) = webc.signature {
            println!("Package signature: {:?}", signature);
        }
    }
    Ok(Version::V2) => {
        let webc = OwnedReader::parse(bytes).unwrap();
        let index = webc.index();
        let signature = &index.signature;
        if !signature.is_none() {
            println!("Package signature: {signature:?}");
        }
    }
    Ok(other) => todo!("Unsupported version, {other}"),
    Err(e) => todo!("An error occurred: {e}"),
}

Re-exports

pub extern crate bytes;

pub extern crate indexmap;

Modules

compatDeprecated

A compatibility layer for dealing with different versions of the WEBC binary format.

metadata

Package metadata.

v1

Parsing code for v1 of the WEBC format.

v2

Parsing code for v2 of the WEBC format.

wasmer_package

Load a Wasmer package from disk.

Structs

Container

A version-agnostic read-only WEBC container.

PathSegment

a cheaply cloneable path segment (i.e. the path, to, and file.txt in path/to/file.txt).

PathSegments

A series of PathSegments specifying the absolute path to something.

Version

A WEBC file’s version number.

Volume

A WEBC volume.

Enums

DetectError

An error returned by detect().

Metadata

Metadata describing the properties of a file or directory.

PathSegmentError

An error that may occur while parsing a PathSegment or PathSegments.

VolumeError

Errors that may occur when doing Volume operations.

Constants

MAGIC

File identification bytes stored at the beginning of the file.

Traits

ToPathSegments

Convert something into PathSegments.

Functions

detect

Check whether something looks like a valid WEBC file and extract its version number.

is_webc

Check if the provided item looks like a WEBC file.

Type Definitions

Magic

The type for MAGIC.