Module webc::wasmer_package

Load a Wasmer package from disk.

Structs

Package

A Wasmer package that will be lazily loaded from disk.

Volume

A lazily loaded volume in a Wasmer package.

Enums

ManifestError

Errors that may occur when converting from a wasmer_toml::Manifest to a crate::metadata::Manifest.

Strictness

The strictness to use when working with a crate::wasmer_package::Package.

WasmerPackageError

Errors that may occur while loading a Wasmer package from disk.